tests/justhtml-tests/table_recovery.dat

#data
<table><input id="x" class="y" type="hidden"></table>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,45): unexpected-hidden-input-in-table

#document
| <html>
|   <head>
|   <body>
|     <table>
|       <input>
|         class="y"
|         id="x"
|         type="hidden"

#data
<table><caption><!--comment--></caption></table>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <!-- comment -->

#data
<table><caption><table></table></caption></table>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,23): unexpected-start-tag-implies-end-tag
(1,23): unexpected-start-tag-implies-end-tag
(1,41): unexpected-end-tag
(1,49): unexpected-end-tag

#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <table>

#data
<table>text</table>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,11): unexpected-character-implies-table-voodoo

#document
| <html>
|   <head>
|   <body>
|     "text"
|     <table>

#data
<table><td><!-- comment in cell --></td></table>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,11): unexpected-cell-in-table-body

#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <!--  comment in cell  -->

#data
<table><colgroup></p>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,18): unexpected-end-tag

#document
| <html>
|   <head>
|   <body>
|     <p>
|     <table>
|       <colgroup>


#data
<table><form><input>x
#errors
#document
| <html>
|   <head>
|   <body>
|     <input>
|     "x"
|     <table>
|       <form>

#data
<table><colgroup></colgroup><colgroup></colgroup></table>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
|       <colgroup>

#data
<table><c><col>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <c>
|     <table>
|       <colgroup>
|         <col>

#data
<table><caption>x<table><tr><td>nested</td></tr></table></caption></table>
#errors
(1,7): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "x"
|         <table>
|           <tbody>
|             <tr>
|               <td>
|                 "nested"

#data
<table><a><th></th></br>
#errors
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <a>
|       <br>
|     <table>
|       <tbody>
|         <tr>
|           <th>

#data
<table></> </l>\x0a<
#errors
#document
| <html>
|   <head>
|   <body>
|     "
<"
|     <table>
|       " "

#data
<table></>\x0a<
#errors
#document
| <html>
|   <head>
|   <body>
|     "
<"
|     <table>

#data
<div><strong></div><table><colgroup>H
#errors
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <strong>
|     <strong>
|       "H"
|     <table>
|       <colgroup>
